requirements.txt
python-telegram-bot==20.7
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.8.6
python-dotenv==1.0.1
pytz==2024.1
